Switch to Droid Sans, a more widely available font than Helvetica Neue
Helvetica Neue is a commercial font, while Droid is available under an
open-source license and included with TeX Live's full distribution by
default (the "droid" package may be installed from CTAN otherwise).
Using filename-based font loading means the bundled font will be loaded
instead of requiring it to be installed on the system.