Containerized web service deployment using Docker
Useful in environments where Docker is easily available but the Perl
ecosystem isn't generally installed.

After building the Docker image locally with:

    docker build -t lab-labels .    # or `make docker-image`

an ephemeral container can be started with:

    docker run --rm --publish 5000:80 lab-labels

Adjust the port mapping as you wish.  During development, a modified
version of the app can be overlaid into the container image like so:

    docker run --rm --publish 5000:80 --volume $PWD:/src lab-labels