Code 128 barcode-generating macro for TeX
Sourced from https://ctan.org/tex-archive/macros/generic/code128,
available under the GPL.

It seems a bit crazy to be generating barcodes with a typesetting
program, but here I am.  It seems to work just fine, and is an expedient
way to get barcodes on these labels!

I'm committing this separately so that I can make my small tweaks to the
macro clear in the next commit.  After tweaking, it will be used to add
barcode generation for labels.