Add support for DTCR-1000 labels on LabelWriter
Printing a few labels at a time is easier and less wasteful on a label
printer than on a sheet of labels fed into a label printer. I opted to
use plain (Xe)TeX instead of LaTeX to have simpler direct control over
what is rendered and where it goes on the page. Printing the PDF output
on a LabelWriter 450 required adding a custom paper size in macOS's
print settings that matches the spec'd size of the label.

I'm not, for now, supporting the barcode functionality Tom Sibley added,
since it's not in use in our group and I'd possibly have to do some
extra futzing around to make loading the macro file work, and I am going
to defer that for the sake of getting the printer set up in the lab
instead of sitting on my desk.